--- src/app/components/AboutPage/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-function AboutPage() {
-  return (
-    <div>
-      <h1>About Page</h1>
-      <Link to="/">Go to HomePage</Link>
-    </div>
-  );
-}
-
-export default AboutPage;
