+++ src/app/components/HomePage/style.scss
