+++ src/app/components/InfoPage/style.scss
