+++ src/app/components/InfoRulesPage/style.scss
