+++ src/app/components/NavBar/style.scss
