+++ src/app/components/Page404/style.scss
